Elasticsearch/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
Django==3.1.3
django-elasticsearch-dsl==7.1.4
djangorestframework==3.12.2
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
